use double in evaluation. some dirty tricks with double precision.

+++ MathEngine/Context/IDataContext.cs
@@ -4,6 +4,6 @@
     {
         IDataContext InnerDataContext { get; }
 
-        decimal ResolveVariable(string name);
+        double ResolveVariable(string name);
     }
 }
