+++ MathEngine/Expressions/Expression.cs
@@ -6,11 +6,11 @@ namespace MathEngine.Expressions
     [DebuggerDisplay("{GetDisplayString()}")]
     public abstract class Expression
     {
-        public abstract decimal Evaluate(IDataContext dataContext);
+        public abstract double Evaluate(IDataContext dataContext);
 
         public abstract string GetDisplayString();
 
-        public static decimal Evaluate(Expression exp, IDataContext context = null)
+        public static double Evaluate(Expression exp, IDataContext context = null)
         {
             return exp.Evaluate(context ?? DefaultContext.Instance);
         }
